+++ pupy/network/lib/transports/cryptoutils/__init__.py
@@ -32,11 +32,13 @@ except ImportError as e:
     def hmac_sha256_digest(key, msg):
         return hmac.new(key, msg, hashlib.sha256).digest()
 
+from .ecpv import ECPV
+
 
 __all__ = (
     append_PKCS7_padding, strip_PKCS7_padding,
     NewAESCipher, get_random,
     SHA1, SHA256, SHA384, SHA3_256, SHA3_512,
     AES_MODE_CTR, AES_MODE_CFB, AES_MODE_CBC,
-    hmac_sha256_digest, AES_BLOCK_SIZE, RC4
+    hmac_sha256_digest, AES_BLOCK_SIZE, RC4, ECPV
 )
